pyproject.toml
[tool.poetry]
name = "physiognomy-classification"
version = "0.1.0"
description = ""
authors = ["Vsevolod <wsewolod@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = {version = "^2.1.1+cu121", source = "pytorch"}
hydra-core = "^1.3.2"
lightning = "^2.1.2"
einops = "^0.7.0"
wandb = "^0.16.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
dlib = "^19.24.4"
pillow = "^10.3.0"
scikit-image = "^0.23.2"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
